+++ hw12_13_14_15_calendar/cmd/scheduler/scheduler.go
@@ -9,7 +9,7 @@ import (
 	_ "github.com/jackc/pgx/stdlib"
 	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/config"
 	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/logger"
-	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/scheduler"
+	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/scheduler/rmqsc"
 )
 
 var configSchedulerFile string
@@ -31,7 +31,7 @@ func main() {
 		return
 	}
 
-	sc := scheduler.NewScheduler(config.Scheduler, storage, logg)
+	sc := rmqsc.NewScheduler(config.Scheduler, storage, logg)
 	defer sc.Cancel()
 
 	ctx, cancel := signal.NotifyContext(context.Background(),
