+++ hw12_13_14_15_calendar/internal/scheduler/sheduler.go
@@ -1,0 +1,10 @@
+package scheduler
+
+import (
+	"context"
+)
+
+type Scheduler interface {
+	Run(context.Context)
+	Cancel()
+}
